+++ Changelog.md
@@ -1,5 +1,6 @@
 # 0.0.10
-+ Added `Feature.from_value` function for creating `Feature` instances from plain Python values.
++ Added `Feature.from_value` class method for creating `Feature` instances from plain Python values.
++ Added `Feature.to_value` method for converting a `Feature` instance to a plain Python value.
 + Removed `features` argument in `Prediction.create` method. Use `inputs` kwargs instead.
 + Removed `FeatureInput` class.